Move common tests stuff in testsUtils
