+++ src/mutations/__tests__/mutationsComputed.test.js
@@ -1,27 +1,10 @@
 import React from 'react'
 import { rj } from '../../rocketjump'
-import rjMiddleware from '../../rjMiddleware'
 import useRj from '../../hooks/useRj'
 import { renderHook, act } from '@testing-library/react-hooks'
-import createSagaMiddleware from 'redux-saga'
-import { createStore, applyMiddleware, combineReducers } from 'redux'
+import { combineReducers } from 'redux'
 import { Provider } from 'react-redux'
-
-const createRealStoreWithSagaAndReducer = (saga, reducer, preloadedState) => {
-  const sagaMiddleware = createSagaMiddleware()
-  const actions = []
-  const actionLogMiddleware = store => next => action => {
-    actions.push(action)
-    return next(action)
-  }
-  const store = createStore(
-    reducer,
-    preloadedState,
-    applyMiddleware(sagaMiddleware, rjMiddleware, actionLogMiddleware)
-  )
-  sagaMiddleware.run(saga)
-  return [store, actions]
-}
+import { createRealStoreWithSagaAndReducer } from '../../testUtils'
 
 describe('RJ mutations computed', () => {
   it('should work as expected without break ... for now', async () => {
